# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: IncidentIQ - Group Incident Reporting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups/{groupSlug}/members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all members of a group
      parameters:
        - $ref: "#/components/parameters/groupSlug"
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupMember"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a member of this group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupSlug}/members/invite:
    post:
      operationId: inviteMember
      tags: [members]
      summary: Invite a user to the group by email (admin only)
      parameters:
        - $ref: "#/components/parameters/groupSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberRequest"
      responses:
        "200":
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin of this group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupSlug}/members/join-requests:
    get:
      operationId: listJoinRequests
      tags: [members]
      summary: List pending join requests (admin only)
      parameters:
        - $ref: "#/components/parameters/groupSlug"
      responses:
        "200":
          description: Pending join requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JoinRequest"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin of this group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupSlug}/members/join-requests/{requestId}:
    post:
      operationId: respondToJoinRequest
      tags: [members]
      summary: Approve or reject a join request (admin only)
      parameters:
        - $ref: "#/components/parameters/groupSlug"
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRequestResponse"
      responses:
        "200":
          description: Request actioned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin of this group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupSlug}/members/{userId}:
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a member's role, permissions, or responder status (admin only)
      parameters:
        - $ref: "#/components/parameters/groupSlug"
        - $ref: "#/components/parameters/userId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRequest"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMember"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin of this group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: removeMember
      tags: [members]
      summary: Remove a member from the group (admin only)
      parameters:
        - $ref: "#/components/parameters/groupSlug"
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin of this group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Incident Types ─────────────────────────────────────────────────────────
...
  /push/subscribe:
    post:
      operationId: subscribePush
      tags: [push]
      summary: Register a Web Push subscription for the current user's device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscribeRequest"
      responses:
        "201":
          description: Push subscription registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /push/unsubscribe:
    post:
      operationId: unsubscribePush
      tags: [push]
      summary: Remove a Web Push subscription for the current user's device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushUnsubscribeRequest"
      responses:
        "200":
          description: Subscription removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Billing ───────────────────────────────────────────────────────────────
...
components:
  parameters:
    groupSlug:
      name: groupSlug
      in: path
      required: true
      schema:
        type: string
      description: The unique slug identifying the group
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
      description: The user's ID
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ─────────────────────────────────────────────────────────────────
...
    MemberPermissions:
      type: object
      properties:
        canReceiveNotifications:
          type: boolean
        canViewDashboard:
          type: boolean
        canActionReports:
          type: boolean
        canFileReports:
          type: boolean
      required: [canReceiveNotifications, canViewDashboard, canActionReports, canFileReports]
    GroupMember:
      type: object
      properties:
        userId:
          type: string
        name:
          type: string
        email:
          type: string
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, responder, member]
        roleTitle:
          type: string
          nullable: true
        permissions:
          $ref: "#/components/schemas/MemberPermissions"
        isActive:
          type: boolean
        joinedAt:
          type: string
          format: date-time
      required: [userId, name, email, role, permissions, isActive, joinedAt]
    InviteMemberRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [member, responder, admin]
          default: member
      required: [email]
    JoinRequest:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        email:
          type: string
        requestedAt:
          type: string
          format: date-time
      required: [id, userId, name, email, requestedAt]
    JoinRequestResponse:
      type: object
      properties:
        action:
          type: string
          enum: [approve, reject]
      required: [action]
    UpdateMemberRequest:
      type: object
      properties:
        role:
          type: string
          enum: [admin, responder, member]
        roleTitle:
          type: string
          nullable: true
        isActive:
          type: boolean
        permissions:
          $ref: "#/components/schemas/MemberPermissions"

    # ── Incident Types ────────────────────────────────────────────────────────
...
    PushSubscribeRequest:
      type: object
      properties:
        endpoint:
          type: string
        p256dh:
          type: string
        auth:
          type: string
        userAgent:
          type: string
          nullable: true
      required: [endpoint, p256dh, auth]
    PushUnsubscribeRequest:
      type: object
      properties:
        endpoint:
          type: string
      required: [endpoint]

    # ── Billing ───────────────────────────────────────────────────────────────
...