# lib/api-zod/src/generated/types/pushSubscribeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IncidentIQ - Group Incident Reporting Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface PushSubscribeRequest {
  endpoint: string;
  p256dh: string;
  auth: string;
  userAgent?: string | null;
}